requirements.txt
# requirements.txt für das Divergenzen-Projekt
pandas
numpy
matplotlib
yfinance
oandapyV20
questionary
pyyaml
